feat : isRunning_ 초기값 false → Run() 루프 진입 불가 버그 수정, 패킷 길이 Uint16 적용

+++ DUBUFrameWork/Client.cpp
@@ -31,15 +31,15 @@ bool DUBU::Client::Dispatch()
 	return true;
 }
 
-void DUBU::Client::SendTo(Uint8* ptr, Int32 size)
+void DUBU::Client::SendTo(Uint8* ptr, Uint16 size)
 {
 	rudpSocket_->SendTo(rudpSocket_->GetSockAddr(), ptr, size);
 }
 
-void DUBU::Client::OnRecvFrom(const SOCKADDR_IN& addr, Uint8* ptr, Int32 size)
+void DUBU::Client::OnRecvFrom(const SOCKADDR_IN& addr, Uint8* ptr, Uint16 size)
 {
 }
 
-void DUBU::Client::OnSendTo(const SOCKADDR_IN& addr, Uint8* ptr, Int32 size)
+void DUBU::Client::OnSendTo(const SOCKADDR_IN& addr, Uint8* ptr, Uint16 size)
 {
 }
